[TI Clang Migration][Phase 1]
[TI CLANG Migration][PDK-9370]Enable build for R5 core

- Initial commit to port compiler to TI CLNAG
- able to compile few files/.asm


[TI CLANG Migration]After compiling 'csl' with some warnings suppressed


[TI CLANG Migration][PDK-9551]OSAL updates

- This commit also includes some changes for build, board, sciclient, GPIO and UART


[TI CLANG Migration][OSAL][SBL][Board]


[TI Clang Migration][PDK-9370][PDK-9550][PDK-9551]

- Additional OSAL changes
- freeRTOS changes
- Liker command file changes for freeRTOS


[TI CLANG Migration][PDK-9554] DSS Driver changes - Patch 2

- OSAL changes for TIRTOS
    - these are enough as auto-generated files by XDS tools also throws errors
- includes '.gitignore' changes to ignore '*.d', '*.o'


[PDK-9555][TI CLANG Migration] R5 Drivers Updates and Testing: IPC, VISS, LDC, NF, SDE, MSC, DOF: Patch-1

- IPC driver changes
- Additional warnings ignored


[TI Clang Migration][Board]Porting to Clang

- increased OCMC memory size to 1MB for J721E for /packages/ti/board/utils/uniflash/target/soc/k3/linker_j7.cmd


[TI Clang Migration][SBL]Porting to Clang


[TI Clang Migration][SBL]Porting to Clang patch-2

- Increased memory section size for SBL OCM section, took the extra space from reserved 'OCMRAM_SBL_RESERVED_CUST_BOOT' section
- check if this is OK?


[TI Clang Migration][Build]Added options for additional warning suppression


[TI Clang Migration][Build]Disabled mcu core for TIRTOS build

- Disabled NIMU as it is only built for TIRTOS

+++ packages/ti/board/diag/current_monitor/src/current_monitor_test.h
@@ -118,7 +118,7 @@ extern "C" {
  *  \brief Structure defining INA Device calibration parameters.
  */
 typedef struct  CalibrationParams{
-    uint8_t maxRShunt;
+    float maxRShunt;
     float shuntVolLsb;
     float busVolLsb;
     float powerLsb;
